messer: diskstats: fix calculation of w_await and r_await
See improved commentary.

The metric really is "the total time for all I/O
operations summed, divided by the number of I/O
operations completed."

I used the merged I/O operation count for the divisor,
but user space count is what matters.

The average response time of the read/write operations,
in milliseconds. This is the end-to-end response time,
including time spent in the queue. It is the response
time that the application making I/O requests sees,
not the response time of the physical disk underlying
the block device.