messer: open()/close() file(s) upon every sample
This refactors writing individual samples to the
output file(s). Writing an individual sample is
now a much more complex operation especially on
the HDF5 file because for every sample the file
is opened newly (in append mode), and then properly
closed. This is to help minimizing the risk for
data corruption. Writing a sample now takes about
10 times as long as before (10-20 ms on my system
instead of 1 ms) but that's OK!